requirements.txt
twython==3.6.0
markovify==0.7.0
numpy==1.11.3
Pillow==5.0.0
